Refactor Authentication Logic

+++ server/Controllers/AuthController.js
@@ -4,7 +4,15 @@ const { createSecretToken } = require('../util/SecretToken');
 const bcrypt = require('bcrypt');
 const jwt = require('jsonwebtoken');
 
-module.exports.Signup = async (req, res, next) => {
+const setCookie = (res, name, value, httpOnly = true) => {
+  res.cookie(name, value, {
+    httpOnly: httpOnly,
+    secure: true,
+    expires: new Date(Date.now() + 1000 * 60 * 60 * 24 * 30 * 6),
+  });
+};
+
+module.exports.Signup = async (req, res) => {
   try {
     const { email, password, name, createdAt } = req.body;
     const existingUser = await User.findOne({ email });
@@ -13,22 +21,17 @@ module.exports.Signup = async (req, res, next) => {
     }
     const user = await User.create({ email, password, name, createdAt });
     const token = createSecretToken(user._id);
-    res.cookie('token', token, {
-      domain: process.env.DOMAIN,
-      httpOnly: true,
-      secure: true,
-      expires: new Date(Date.now() + 1000 * 60 * 60 * 24 * 30 * 6),
-    });
+    setCookie(res, 'token', token);
+    setCookie(res, 'name', user.name, false);
     res
       .status(201)
       .json({ name: user.name, message: 'Account created successfully!', success: true });
-    next();
   } catch (error) {
     console.error(error);
   }
 };
 
-module.exports.Login = async (req, res, next) => {
+module.exports.Login = async (req, res) => {
   try {
     const { email, password } = req.body;
     if (!email || !password) {
@@ -43,45 +46,42 @@ module.exports.Login = async (req, res, next) => {
       return res.json({ message: 'Incorrect password or email.' });
     }
     const token = createSecretToken(user._id);
-    res.cookie('token', token, {
-      domain: process.env.DOMAIN,
-      httpOnly: true,
-      secure: true,
-      expires: new Date(Date.now() + 1000 * 60 * 60 * 24 * 30 * 6),
-    });
+    setCookie(res, 'token', token);
+    setCookie(res, 'name', user.name, false);
     res.status(201).json({ name: user.name, message: 'Logged in successfully!', success: true });
-    next();
   } catch (error) {
     console.error(error);
   }
 };
 
-module.exports.Logout = async (req, res, next) => {
+module.exports.Logout = async (req, res) => {
   try {
+    res.clearCookie('name', { domain: process.env.DOMAIN });
     res.clearCookie('token', { domain: process.env.DOMAIN });
     res.status(200).json({ message: 'Logged out successfully!' });
-    next();
   } catch (error) {
     console.error(error);
   }
 };
 
-module.exports.checkAuth = async (req, res, next) => {
+module.exports.getPersonalDetails = async (req, res) => {
   try {
     const token = req.cookies.token;
     if (!token) {
-      return res.status(401).json({ message: 'Unauthorized' });
+      return res.status(401).json({ message: 'Unauthorized request.' });
     }
-
     jwt.verify(token, process.env.TOKEN_KEY, async (err, data) => {
       if (err) {
-        res.status(401).json({ error: 'Request is not authorized' });
+        res.status(401).json({ error: 'Unauthorized request.' });
       } else {
         const user = await User.findById(data.id);
         if (user) {
+          const token = createSecretToken(user._id);
+          setCookie(res, 'token', token);
+          setCookie(res, 'name', user.name, false);
           return res.status(200).json({ name: user.name, email: user.email });
         } else {
-          return res.status(401).json({ error: 'Request is not authorized' });
+          return res.status(401).json({ error: 'Unauthorized request.' });
         }
       }
     });
@@ -90,28 +90,34 @@ module.exports.checkAuth = async (req, res, next) => {
   }
 };
 
-module.exports.updatePersonalDetails = async (req, res, next) => {
+module.exports.updatePersonalDetails = async (req, res) => {
   try {
     const { name, email } = req.body;
     const token = req.cookies.token;
     if (!token) {
-      return res.status(401).json({ message: 'Unauthorized' });
+      return res.status(401).json({ message: 'Unauthorized request.' });
     }
-
     jwt.verify(token, process.env.TOKEN_KEY, async (err, data) => {
       if (err) {
-        res.status(401).json({ error: 'Request is not authorized' });
+        res.status(401).json({ error: 'Unauthorized request.' });
       } else {
-        const existingUser = await User.findOne({ email });
-        if (existingUser && existingUser._id.toString() !== data.id) {
-          return res.json({ success: false, message: 'This email is already in use!' });
+        const user = await User.findById(data.id);
+        if (user) {
+          const existingUser = await User.findOne({ email });
+          if (existingUser && existingUser._id.toString() !== data.id) {
+            return res.json({ success: false, message: 'This email is already in use!' });
+          }
+          const user = await User.findByIdAndUpdate(
+            data.id,
+            { $set: { name, email } },
+            { new: true }
+          );
+          res
+            .status(200)
+            .json({ success: true, message: 'Personal details updated successfully!' });
+        } else {
+          return res.status(401).json({ error: 'Unauthorized request.' });
         }
-        const user = await User.findByIdAndUpdate(
-          data.id,
-          { $set: { name, email } },
-          { new: true }
-        );
-        res.status(200).json({ success: true, message: 'Personal details updated successfully!' });
       }
     });
   } catch (error) {
@@ -119,30 +125,33 @@ module.exports.updatePersonalDetails = async (req, res, next) => {
   }
 };
 
-module.exports.updatePassword = async (req, res, next) => {
+module.exports.updatePassword = async (req, res) => {
   try {
     const { currentPassword, newPassword } = req.body;
     const token = req.cookies.token;
     if (!token) {
-      return res.status(401).json({ message: 'Unauthorized' });
+      return res.status(401).json({ message: 'Unauthorized request.' });
     }
-
     jwt.verify(token, process.env.TOKEN_KEY, async (err, data) => {
       if (err) {
-        res.status(401).json({ error: 'Request is not authorized' });
+        res.status(401).json({ error: 'Unauthorized request.' });
       } else {
         let user = await User.findById(data.id);
-        const auth = await bcrypt.compare(currentPassword, user.password);
-        if (!auth) {
-          return res.json({ success: false, message: 'Incorrect password.' });
+        if (user) {
+          const auth = await bcrypt.compare(currentPassword, user.password);
+          if (!auth) {
+            return res.json({ success: false, message: 'Incorrect password.' });
+          }
+          const hashedPassword = await bcrypt.hash(newPassword, 12);
+          user = await User.findByIdAndUpdate(
+            data.id,
+            { $set: { password: hashedPassword } },
+            { new: true }
+          );
+          res.status(200).json({ success: true, message: 'Password updated successfully!' });
+        } else {
+          return res.status(401).json({ error: 'Unauthorized request.' });
         }
-        const hashedPassword = await bcrypt.hash(newPassword, 12);
-        user = await User.findByIdAndUpdate(
-          data.id,
-          { $set: { password: hashedPassword } },
-          { new: true }
-        );
-        res.status(200).json({ success: true, message: 'Password updated successfully!' });
       }
     });
   } catch (error) {
@@ -150,24 +159,29 @@ module.exports.updatePassword = async (req, res, next) => {
   }
 };
 
-module.exports.deleteAccount = async (req, res, next) => {
+module.exports.deleteAccount = async (req, res) => {
   try {
     const token = req.cookies.token;
     if (!token) {
-      return res.status(401).json({ message: 'Unauthorized' });
+      return res.status(401).json({ message: 'Unauthorized request.' });
     }
-
     jwt.verify(token, process.env.TOKEN_KEY, async (err, data) => {
       if (err) {
-        res.status(401).json({ error: 'Request is not authorized' });
+        res.status(401).json({ error: 'Unauthorized request.' });
       } else {
-        const coffee = await Coffee.find({ userId: data.id });
-        if (coffee.length > 0) {
-          await Coffee.deleteMany({ userId: data.id });
+        let user = await User.findById(data.id);
+        if (user) {
+          await User.findByIdAndDelete(data.id);
+          const coffee = await Coffee.find({ userId: data.id });
+          if (coffee.length > 0) {
+            await Coffee.deleteMany({ userId: data.id });
+          }
+          res.clearCookie('name', { domain: process.env.DOMAIN });
+          res.clearCookie('token', { domain: process.env.DOMAIN });
+          res.status(200).json({ success: true, message: 'Account deleted successfully!' });
+        } else {
+          return res.status(401).json({ error: 'Unauthorized request.' });
         }
-        await User.findByIdAndDelete(data.id);
-        res.clearCookie('token', { domain: process.env.DOMAIN });
-        res.status(200).json({ success: true, message: 'Account deleted successfully!' });
       }
     });
   } catch (error) {
